# 3week/docs/API/api.yaml
openapi: 3.0.0
info:
  title: E-Commerce Order API
  description: e-커머스 상품 주문 서비스 API 문서
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 로컬 서버
...
paths:
...
  /coupons/{userId}:
    post:
      summary: 쿠폰 발급
      tags:
        - COUPON
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                couponId:
                  type: integer
      responses:
        200:
          description: 쿠폰 발급 성공
        400:
          description: 쿠폰 수량 부족 또는 유효기간 만료
    get:
      summary: 보유하는 쿠폰 목록
      tags:
        - COUPON
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  couponId:
                    type: integer
                    description: 쿠폰 아이디
                  couponNumber:
                    type: string
                    description: 쿠폰번호
                  type:
                    type: string
                    description: 할인 종류(정액, 정률)
                  rate:
                    type: integer
                    description: 할인율
                  discountPrice:
                    type: integer
                    description: 할인 금액
                  isUsed:
                    type: boolean
                    description: 쿠폰 사용 여부
...